requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0
python-dotenv==1.0.0
python-multipart==0.0.6

# For Excel support
openpyxl==3.10.10
odfpy==1.4.1

# CORS and security
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6

# Scheduled tasks
APScheduler==3.10.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.12.0
flake8==6.1.0
